RouteView - refresh on Goto
bizarely
only when goto shifts levels to select the next node
the panel needs to be invalidated (or a mouseover on
the panel will do the same). But this happens only to
RouteView (not TopRouteView(Route)). Nor does it
happen when a node is selected by Goto linked node eg.